ipipe: x86: Fix flags corruption in arch_mangle_irq_bits on 64-bit systems
Due to unwanted sign extension, we set all high bits in the mangled
flags on x86-64. Harmless on current CPUs, but that may change one day.

+++ arch/x86/include/asm/irqflags.h
@@ -143,7 +143,7 @@ static inline void halt(void)
 /* Merge virtual+real interrupt mask bits into a single word. */
 static inline unsigned long arch_mangle_irq_bits(int virt, unsigned long real)
 {
-	return (real & ~(1L << 31)) | ((virt != 0) << 31);
+	return (real & ~(1L << 31)) | ((unsigned long)(virt != 0) << 31);
 }
 
 /* Converse operation of arch_mangle_irq_bits() */
